rename join function to twJoin internally
